+++ drivers/can/CMakeLists.txt
@@ -2,6 +2,7 @@
 
 zephyr_sources_ifdef(CONFIG_CAN          can_common.c)
 zephyr_sources_ifdef(CONFIG_CAN_LOOPBACK can_loopback.c)
-zephyr_sources_ifdef(CONFIG_CAN_MCP2515  mcp2515.c)
-zephyr_sources_ifdef(CONFIG_CAN_STM32    stm32_can.c)
+zephyr_sources_ifdef(CONFIG_CAN_MCP2515  can_mcp2515.c)
+zephyr_sources_ifdef(CONFIG_CAN_STM32    can_stm32.c)
+
 zephyr_sources_ifdef(CONFIG_USERSPACE    can_handlers.c)
